requirements.txt
groq
python-dotenv
streamlit
pandas
pytest
tabulate
